# performance improvements for `any` and `all`