* Help ensure vectorization of `any` and `all` for homogeneous tuples of `Bool`, by avoiding bounds-checking and avoiding short-circuiting. Inspired by #55673, but more general by relying on a loop instead of on recursion, thus being performant for any input length.
* Delete a bunch of now-useless methods.
    * *eight* methods of `all` are deleted
    * four methods of `any` are deleted

Closes #55673