@nsajko 2025-01-18
Here's the `code_llvm`:

```julia-repl
julia> code_llvm(all, Tuple{Tuple{Vararg{Bool, 32}}}; debuginfo=:none)
; Function Signature: all(NTuple{32, Bool})
define i8 @julia_all_2401(ptr nocapture noundef nonnull readonly align 1 dereferenceable(32) %"itr::Tuple") #0 {
top:
  %0 = load <32 x i8>, ptr %"itr::Tuple", align 1
  %1 = icmp eq <32 x i8> %0, zeroinitializer
  %2 = bitcast <32 x i1> %1 to i32
  %3 = icmp eq i32 %2, 0
  %4 = zext i1 %3 to i8
  ret i8 %4
}

julia> code_llvm(any, Tuple{Tuple{Vararg{Bool, 32}}}; debuginfo=:none)
; Function Signature: any(NTuple{32, Bool})
define i8 @julia_any_2403(ptr nocapture noundef nonnull readonly align 1 dereferenceable(32) %"itr::Tuple") #0 {
top:
  %0 = load <32 x i8>, ptr %"itr::Tuple", align 1
  %1 = call i8 @llvm.vector.reduce.or.v32i8(<32 x i8> %0)
  %2 = icmp ne i8 %1, 0
  %3 = zext i1 %2 to i8
  ret i8 %3
}

julia> code_llvm(all, Tuple{Tuple{Vararg{Bool, 64}}}; debuginfo=:none)
; Function Signature: all(NTuple{64, Bool})
define i8 @julia_all_2405(ptr nocapture noundef nonnull readonly align 1 dereferenceable(64) %"itr::Tuple") #0 {
top:
  %wide.load = load <32 x i8>, ptr %"itr::Tuple", align 1
  %0 = icmp ne <32 x i8> %wide.load, zeroinitializer
  %1 = getelementptr inbounds i8, ptr %"itr::Tuple", i64 32
  %wide.load.1 = load <32 x i8>, ptr %1, align 1
  %2 = icmp ne <32 x i8> %wide.load.1, zeroinitializer
  %3 = and <32 x i1> %0, %2
  %4 = bitcast <32 x i1> %3 to i32
  %5 = icmp eq i32 %4, -1
  %6 = zext i1 %5 to i8
  ret i8 %6
}

julia> code_llvm(any, Tuple{Tuple{Vararg{Bool, 64}}}; debuginfo=:none)
; Function Signature: any(NTuple{64, Bool})
define i8 @julia_any_2407(ptr nocapture noundef nonnull readonly align 1 dereferenceable(64) %"itr::Tuple") #0 {
top:
  %wide.load = load <32 x i8>, ptr %"itr::Tuple", align 1
  %0 = getelementptr inbounds i8, ptr %"itr::Tuple", i64 32
  %wide.load.1 = load <32 x i8>, ptr %0, align 1
  %1 = or <32 x i8> %wide.load, %wide.load.1
  %2 = icmp ne <32 x i8> %1, zeroinitializer
  %3 = bitcast <32 x i1> %2 to i32
  %4 = icmp ne i32 %3, 0
  %5 = zext i1 %4 to i8
  ret i8 %5
}

julia> versioninfo()
Julia Version 1.12.0-DEV.unknown
Commit 319082c (2025-01-18 07:54 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, znver2)
  GC: Built with stock GC
Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
```

So it vectorizes, like in #55673, but it also vectorizes for lengths above 32.