@Seelengrab 2025-01-18
Do you have some benchmarks showing the improved performance?