@nsajko 2025-01-18
## Benchmarking script

```julia
using BenchmarkTools

for f ∈ (all, any)
    println("f: $f")
    for b ∈ (false, true)
        println("  b: $b")
        for l ∈ 32:32:96
            println("    l: $l")
            print("    ")
            @btime ($f)(t) setup=(t = ntuple(Returns($b), $l);)
        end
    end
end
```

## Results

### `master` branch

Commit 17402876ec2 (0 days old master):

```none
f: all
  b: false
    l: 32
      13.392 ns (0 allocations: 0 bytes)
    l: 64
      13.143 ns (0 allocations: 0 bytes)
    l: 96
      18.287 ns (0 allocations: 0 bytes)
  b: true
    l: 32
      24.725 ns (0 allocations: 0 bytes)
    l: 64
      46.582 ns (0 allocations: 0 bytes)
    l: 96
      64.334 ns (0 allocations: 0 bytes)
f: any
  b: false
    l: 32
      29.442 ns (0 allocations: 0 bytes)
    l: 64
      54.097 ns (0 allocations: 0 bytes)
    l: 96
      74.817 ns (0 allocations: 0 bytes)
  b: true
    l: 32
      13.652 ns (0 allocations: 0 bytes)
    l: 64
      13.160 ns (0 allocations: 0 bytes)
    l: 96
      22.520 ns (0 allocations: 0 bytes)
```

### PR branch

nsajko:any_all_vectorized_tuple_bool/a6a6f6b0746 (fork: 2 commits, 0 days):

```none
f: all
  b: false
    l: 32
      13.747 ns (0 allocations: 0 bytes)
    l: 64
      13.908 ns (0 allocations: 0 bytes)
    l: 96
      18.239 ns (0 allocations: 0 bytes)
  b: true
    l: 32
      13.716 ns (0 allocations: 0 bytes)
    l: 64
      13.924 ns (0 allocations: 0 bytes)
    l: 96
      21.446 ns (0 allocations: 0 bytes)
f: any
  b: false
    l: 32
      13.652 ns (0 allocations: 0 bytes)
    l: 64
      17.535 ns (0 allocations: 0 bytes)
    l: 96
      18.269 ns (0 allocations: 0 bytes)
  b: true
    l: 32
      13.652 ns (0 allocations: 0 bytes)
    l: 64
      14.686 ns (0 allocations: 0 bytes)
    l: 96
      21.867 ns (0 allocations: 0 bytes)
```

## Interpretation

Speedups for `(all, true)` and `(any, false)`.