@nsajko 2025-01-18
I think the functions I added might not be necessary if instead this part were improved, by ensuring there's no bounds checking:

https://github.com/JuliaLang/julia/blob/17402876ec235dd66452983a39f8521f9ff62936/base/anyall.jl#L117-L124

Trying that out ...