@nsajko 2025-01-22
Ready for review.

Just for fun, here's what happens at even bigger tuple lengths: the run time drops from microseconds to nanoseconds in some cases. The allocations are, I guess, because tuples seem to get passed by reference when longer than 511 elements (?).

## Benchmarking script

```julia
using BenchmarkTools

for f ∈ (all, any)
    println("f: $f")
    for b ∈ (false, true)
        println("  b: $b")
        for l ∈ (2^11):(2^11):(2^13)
            println("    l: $l")
            print("    ")
            @btime ($f)(t) setup=(t = ntuple(Returns($b), $l);)
        end
    end
end
```

## Results

### `master` branch

Commit 2e6ffbccab2 (1 day old master):

```none
f: all
  b: false
    l: 2048
      18.621 ns (1 allocation: 16 bytes)
    l: 4096
      18.628 ns (1 allocation: 16 bytes)
    l: 6144
      24.967 ns (1 allocation: 16 bytes)
    l: 8192
      39.449 ns (1 allocation: 16 bytes)
  b: true
    l: 2048
      1.088 μs (1 allocation: 16 bytes)
    l: 4096
      1.612 μs (1 allocation: 16 bytes)
    l: 6144
      3.204 μs (1 allocation: 16 bytes)
    l: 8192
      3.212 μs (1 allocation: 16 bytes)
f: any
  b: false
    l: 2048
      1.083 μs (1 allocation: 16 bytes)
    l: 4096
      2.143 μs (1 allocation: 16 bytes)
    l: 6144
      3.200 μs (1 allocation: 16 bytes)
    l: 8192
      4.254 μs (1 allocation: 16 bytes)
  b: true
    l: 2048
      18.681 ns (1 allocation: 16 bytes)
    l: 4096
      18.661 ns (1 allocation: 16 bytes)
    l: 6144
      27.086 ns (1 allocation: 16 bytes)
    l: 8192
      30.621 ns (1 allocation: 16 bytes)
```

### PR branch

any_all_vectorized_tuple_bool/619d80f1d23 (fork: 9 commits, 0 days):

```none
f: all
  b: false
    l: 2048
      42.689 ns (1 allocation: 16 bytes)
    l: 4096
      37.465 ns (1 allocation: 16 bytes)
    l: 6144
      60.798 ns (1 allocation: 16 bytes)
    l: 8192
      69.486 ns (1 allocation: 16 bytes)
  b: true
    l: 2048
      49.109 ns (1 allocation: 16 bytes)
    l: 4096
      37.440 ns (1 allocation: 16 bytes)
    l: 6144
      68.439 ns (1 allocation: 16 bytes)
    l: 8192
      73.108 ns (1 allocation: 16 bytes)
f: any
  b: false
    l: 2048
      32.737 ns (1 allocation: 16 bytes)
    l: 4096
      48.776 ns (1 allocation: 16 bytes)
    l: 6144
      61.788 ns (1 allocation: 16 bytes)
    l: 8192
      71.233 ns (1 allocation: 16 bytes)
  b: true
    l: 2048
      36.217 ns (1 allocation: 16 bytes)
    l: 4096
      48.554 ns (1 allocation: 16 bytes)
    l: 6144
      71.521 ns (1 allocation: 16 bytes)
    l: 8192
      71.202 ns (1 allocation: 16 bytes)
```